% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akin-defunct.R
\name{akin-defunct}
\alias{akin-defunct}
\title{Defunct Functions In Package akin}
\description{
The following functions are defunct. Where available, alternative functions are suggested
at \code{help("akin-defunct")}.
}
\section{\code{common}}{
 this function is defunct. Use \link{fcommon} instead.
}

\keyword{internal}
