% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{common}
\alias{common}
\title{Identify Common Substrings In A Pair Of Strings}
\usage{
common(
  X,
  Y = NULL,
  from,
  to,
  lower = NULL,
  upper = NULL,
  outlist = FALSE,
  rows = 1000,
  wait = 100,
  ...
)
}
\arguments{
\item{X, Y}{character, length = 1 each: a string, such as a protein chain. \code{Y} can be \code{NULL}.}

\item{from, to}{integer length 1 each. Range of string number of characters to identify. When \code{to} is missing from call,
only substrings of number of characters equal to \code{from} are sought}

\item{lower, upper}{integer, length 1 each, default NULL: all combinations. Otherwise, only a subset of all combinations starting
with the \code{lower}-th and ending with the \code{upper}-th combinations. See \link[RcppAlgos:combinatoricsGeneral]{RcppAlgos::comboGeneral}}

\item{outlist}{logical. Default, FALSE, return a character vector of common substrings. Otherwise, return valid
substrings found in each chain}

\item{rows}{integer length 1. Default 1000. The number of rows in each iteration (i.e. combinations) matrix}

\item{wait}{integer length 1. Default 100. Duration of background process polling. Unit: milliseconds.
During this time, the main process is being blocked.}

\item{...}{not used}
}
\value{
A character vector or list of substrings found in both chains. Otherwise, "character(0)". When \code{Y = NULL},
a list of all valid substrings within the \code{from}-\code{to} window.
}
\description{
Checks and identifies substrings that are common in a pair of strings.
}
\details{
Using brute combinatorial approach, this utility splits each of the two \code{chain}s into all possible
combinations of substrings with lengths inside the \code{from}-\code{to} window. Each combination is filtered for
uniqueness and appartenance to its original chain (i.e. for validity). Search time increases severely with the number
of combinations: \code{chain} and substring lengths, and search window width. Slight performance improvement comes
from variations of \code{rows} values as the larger the number of \code{rows}, the smaller the number of iterations.
The default value is a satisfactory starting point. Chains of hundreds of characters, such as long protein chains,
may take hours on some machines. In such cases, \code{rows} can be set to the order of 1e6. During search, the
partitioning of chains runs asynchronously when \code{Y != NULL}.

\code{lower} and \code{upper} limits. Setting any or both of these limits reduces the search time without guaranteeing
the completeness of common substrings list. It helps establishing the existence of \emph{any} common substrings.

NOTE: This utility uses background processing. Check "Security Considerations" in \strong{callr} package documentation.
}
\examples{

if (interactive()) {

# 1. A set of chains

 X = 'alpdxoipoyloiekladxoipoylyl'
 Y = 'kdxoipoylyydxoipoylopldxoipoylac'

# 1.1 Identify all 4-character common substrings
 system.time(a <- common(X, Y, 4))
 print(a)

 # 1.2 Check if substrings in "a" are common
 chain = list(chain1 = 'alpdxoipoyloiekladxoipoylyl',
              chain2 = "kdxoipoylyydxoipoylopldxoipoylac")

 b = sapply(a, findLoc, chain, TRUE, TRUE, TRUE, simplify = FALSE)
 print(b)                                              # a named list


 any(lengths(b) == 0L)                                 # FALSE
 identical(length(a), length(b))                       # TRUE

}

}
\seealso{
\link[RcppAlgos]{comboIter}, \link[RcppAlgos]{comboGeneral}, \link[RcppAlgos]{comboCount}
}
\keyword{Proteomics}
