% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{cover}
\alias{cover}
\title{Calculate The Number Of Combinations Of String Covering}
\usage{
cover(x, wplot = FALSE)
}
\arguments{
\item{x}{integer, length = 1 or character vector, length > 1, or character string}

\item{wplot}{logical, FALSE. When TRUE, displays a plot of combinations versus number of characters in substring}
}
\value{
A prettified named vector showing the total, minimum and maximum number of combinations for respective
string, along with a plot (\code{wplot = TRUE}).
}
\description{
Calculates the number of subset family (\emph{covering}) combinations for common substrings
identified by function \link{fcommon}.
}
\details{
This function helps in deciding the parameters for parallel tasks of function \link{fcommon}. It finds the
total, the minimum and the maximum of number of combinations processed by \link{fcommon} for individual substring covering.

Since at peak workload each worker receives by default 1e6 combinations for processing (see argument \code{...} in \link{fcommon}
documentation), the number of workers in a \link[future]{plan} is effectively decided taking the maximum number of
combinations as reference. The \code{min:max} range relates to workers idling outside the peak: setting a plan with too
many workers would render many of them idle during \link{fcommon}'s parallel run. On another hand, setting it with too few, may
challenge the default workload limit (see argument \code{maxSize} in \link{fcommon} documentation).

When \code{x} is an integer, it must equal either the \code{nchar(string)} or the \code{length(character)}.
}
\examples{

if (interactive()) {

# 1. x is a character

# 1.1 A string
x = 'tyrvvsvltvlhqdwlngkeykck'

a = cover(x)
print(a)

# 1.2 A vector
x = strsplit(x, '')[[1]]

b = cover(x)
print(b)

# 2 An integer

n = length(x)
c = cover(n, TRUE)
print(c)

}


}
\seealso{
\link{fcommon}
}
\keyword{Proteomics}
