% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLONG.R
\name{splitV}
\alias{splitV}
\title{Split Data Vertically Into Subsets}
\usage{
splitV(data, splits)
}
\arguments{
\item{data}{a "data.table" class data frame or convertible to "data.table" class}

\item{splits}{integer, length 1, of value <= \code{ncol(data)}. The number of disjoint
selections (i.e. vertical splits). When \code{splits = 0}, there are no splits}
}
\value{
An exhaustive \link[listenv]{listenv} of disjoint column groups from original data
}
\description{
Splits data into unequal and disjoint groups of columns (i.e. vertical splits)
}
\details{
The smaller the \code{splits} value, the wider the column groups. Column order from
source data is not preserved
}
\examples{

if (interactive()) {

# 1. Split iris data vertically

data(iris)
a = splitV(iris, splits = 3)     # split data in 3 column groups
class(a)                         # listenv, environment
print(as.list(a))                # list

}
}
\keyword{splitting}
