% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/akin-package.R
\docType{package}
\name{akin-package}
\alias{akin}
\alias{akin-package}
\title{Functional Utilities For Data Processing}
\description{
The intent of this package is utilitarian which explains the relatively large number of informative messages
displayed by most functions. Designed for large and very large data files, the package employs data.table,
list environments, sparse matrices, C++ combinatorial methods, and background processing in few places.

Nevertheless, Users could try thread optimization, memoisation or parallel processing, as these techniques were
outside the scope.

The package covers several areas of data processing: subset reading and writing of large data files, data tiling
(horizontal and vertical splitting) - suited for data conversion operations with local as well as global hold
such as, one-hot encoding and scoring - stratified, proportional, random or probabilistic data sampling, data normalization
and thresholding, substring location inside strings e.g. peptides inside protein chains, common substrings identification
and location-tabulation of amino acids, modifications or their associated monoisotopic masses inside modified peptides,
covering various representations of protein mass spectrometry data, with no pretense of exhaustiveness.

Comments and suggestions regarding tricky situations are provided however brief. Examples should be run individually
in the R console. When possible, suppression of messages should be avoided.
}
\author{
\strong{Maintainer}: Dragos Bandur \email{dbandur@sympatico.ca}

Authors:
\itemize{
  \item Dragos Bandur \email{dbandur@sympatico.ca}
}

}
\keyword{internal}
