% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cover.R
\name{cover}
\alias{cover}
\title{Calculate The Number Of Combinations Of String Covering}
\usage{
cover(x, valid. = FALSE, wplot = FALSE)
}
\arguments{
\item{x}{integer, length = 1 or character vector, length > 1, or character string. White space is removed}

\item{valid.}{logical, default, FALSE. When TRUE, valid substrings are returned. Requires x as character or string}

\item{wplot}{logical, FALSE. When TRUE, displays a plot of combinations versus number of characters in string covering}
}
\value{
When \code{valid. = FALSE} a prettified named vector showing the total, minimum and maximum number of
combinations for respective string covering. When \code{valid. = TRUE}, a character vector of valid substrings
having min. 2 characters each. In both cases, a plot is returned when \code{wplot = TRUE}.
}
\description{
Calculates the cardinality of substring family (\emph{the covering}) and returns valid elements.
}
\details{
This function finds the total, the minimum and the maximum number of combinations of individual string
covering or returns a subset of valid substrings (i.e. elements of the string) by retaining even transpositions
with \emph{sequential} values inside the combinations matrix. Very long strings may result in memory allocation
error when \code{valid. = TRUE}. Use \link{fcommon} instead.

NOTE: Herein, the \emph{covering} represents the substring family - of min. 2 characters each - that preserves the
order of characters inside the string. Hence, the use of combinations instead of partial permutations. Valid
substrings form a (lower cardinality) subset having elements in sequence (see Example 3).
}
\examples{

if (interactive()) {

# 1.1 A string
 x = 'tyrvvsvltvlhqdwlngkeykck'

# 1.2 A vector
 y = strsplit(x, '')[[1]]

# 1.3 An integer
 n = length(y)

ll = list(str = x, char = y, int = n)
print(t(
       sapply(ll, cover)
                        ))

# 2. Valid substrings with plot

d = cover(x, TRUE, TRUE)
print(head(d, 30))                         # first 30 substrings

# 3. Valid set versus covering

# The plot shows 3 combinations of 2-character groups while e
# contains 2 sequences of 2 characters each, substring "ac"
# not being a sequence in x.

x = letters[1:3]
e = cover(x, TRUE, TRUE)
print(e)                                   # valid combinations

}


}
\seealso{
\link{fcommon}, \link[RcppAlgos]{comboCount}, \link[RcppAlgos]{comboGeneral}
}
\keyword{Proteomics}
