% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcommon.R
\name{fcommon}
\alias{fcommon}
\title{Fast Identify Common Substrings In A Pair Of Strings}
\usage{
fcommon(x, y, ...)
}
\arguments{
\item{x, y}{character, length = 1 each: a string, such as a protein chain. \code{y} can be missing.}

\item{...}{reserved for internal argument \code{rows} which has a default value of 50}
}
\value{
A closure (see Example) which in turn returns a character vector of common substrings of all lengths. When \code{y}
is missing from call, a character vector of all valid substrings in \code{x}.
}
\description{
Fast version of \link{common}. Experimental.
}
\details{
This utility is experimental and will replace \link{common} in near future. It identifies all-length
substrings common to the pair of strings.
}
\examples{

if (interactive()) {

# 1. A pair of strings

 x = 'alpdxoipoyloiekladxoipoylyl'
 y = 'kdxoipoylyydxoipoylopldxoipoylac'

# 1.1 Equivalent to "common(..., from = 2, to = 9)"
 g = fcommon(x, y)
 system.time( a <- g() )
 print(a)                                              # common substrings set

 }

}
\keyword{Proteomics}
